+++ src/components/Header.js
@@ -1,20 +1,46 @@
-import { signOut } from "firebase/auth";
+import { onAuthStateChanged, signOut } from "firebase/auth";
 import { auth } from "../utils/firebase";
 import { useNavigate } from "react-router-dom";
 import { useSelector } from "react-redux";
+import { useEffect } from "react";
+import { addUser, removeUser } from "../utils/userSlice";
+import { useDispatch } from "react-redux";
 
 const Header = () => {
+  const dispath = useDispatch();
   const navigate = useNavigate();
   const user = useSelector((store) => store.user);
   const signoutHandler = () => {
     signOut(auth)
-      .then(() => {
-        navigate("/");
-      })
+      .then(() => {})
       .catch((error) => {
         console.error(error);
       });
   };
+  useEffect(() => {
+    onAuthStateChanged(auth, (user) => {
+      if (user) {
+        // User is signed in, see docs for a list of available properties
+        // https://firebase.google.com/docs/reference/js/auth.user
+        const { uid, email, displayName, photoURL } = user;
+        dispath(
+          addUser({
+            uid: uid,
+            email: email,
+            displayName: displayName,
+            photoURL: photoURL,
+          })
+        );
+        navigate("/browse")
+        // ...
+      } else {
+        dispath(removeUser());
+        // User is signed out
+        // ...
+        navigate("/")
+      }
+    });
+  }, []);
   return (
     <div className=" absolute px-2 py-2 bg-gradient-to-b from-black z-10 w-full flex justify-between">
       <img
